requirements.txt
python-dotenv
google-generativeai
mcp-core 
simple-salesforce
fastapi
uvicorn[standard]
httpx